🚚 [app.middlewares] => rename file
To match the name of its main class `BrowserMiddleware`, it is not anymore only about caching.
