+++ src/youwol/app/middlewares/__init__.py
@@ -3,6 +3,6 @@ This module gather the middlewares specific to the local youwol servers.
 """
 
 # relative
-from .browser_caching_middleware import *
+from .browser_middleware import *
 from .hybridizer_middleware import *
 from .local_cloud_hybridizers import *
