+++ src/youwol/app/middlewares/browser_middleware.py
